Move helper functions into `utils.js`.
