+++ lib/profileAgents.js
@@ -8,17 +8,18 @@ import * as meterClient from './meterClient.js';
 import * as profileMeters from './profileMeters.js';
 import * as utils from './utils.js';
 import * as zcaps from './zcaps.js';
-import {
-  AsymmetricKey, CapabilityAgent, Hmac, KeyAgreementKey, KmsClient
-} from '@digitalbazaar/webkms-client';
-import {EdvClient, EdvDocument} from '@digitalbazaar/edv-client';
+import {CapabilityAgent, KmsClient} from '@digitalbazaar/webkms-client';
 import assert from 'assert-plus';
+import {EdvClient} from '@digitalbazaar/edv-client';
 import {httpsAgent} from '@bedrock/https-agent';
 
 const {
   assertKeystoreOptions,
   createCapabilityAgent,
   getCollection,
+  getEdvConfig,
+  getEdvDocument,
+  getProfileSigner,
   getPublicAliasTemplate
 } = utils;
 const {config, util: {BedrockError}} = bedrock;
@@ -647,7 +648,7 @@ async function _refreshProfileAgentZcaps({profileAgentRecords}) {
         if(expTime - now <= zcapRefreshThreshold) {
           if(zcapName === 'userDocument') {
             const kmsClient = new KmsClient({httpsAgent});
-            const profileSigner = await _getProfileSigner({
+            const profileSigner = await getProfileSigner({
               kmsClient, profileAgentRecord: record, zcaps
             });
             const docUrl = new URL(zcap.invocationTarget);
@@ -655,10 +656,10 @@ async function _refreshProfileAgentZcaps({profileAgentRecords}) {
               `${docUrl.protocol}//${docUrl.hostname}:${docUrl.port}` +
               `${docUrl.pathname.split('/').slice(0, 3).join('/')}`;
             const edvClient = new EdvClient({id: edvId, httpsAgent});
-            const edvConfig = await _getEdvConfig({edvClient, profileSigner});
+            const edvConfig = await getEdvConfig({edvClient, profileSigner});
             const docId = zcap.invocationTarget.split('/').pop();
             // get profile agent user doc
-            const profileAgentUserDoc = await _getEdvDocument({
+            const profileAgentUserDoc = await getEdvDocument({
               docId, edvConfig, edvClient, kmsClient, profileSigner
             });
             const updatedProfileAgentUserDoc =
@@ -704,57 +705,6 @@ async function _refreshProfileAgentZcaps({profileAgentRecords}) {
     }
   }
   return zcapRefreshed;
-}
-
-async function _getEdvConfig({edvClient, profileSigner} = {}) {
-  return edvClient.getConfig({
-    invocationSigner: profileSigner,
-  });
-}
-
-async function _getEdvDocument({
-  docId, edvConfig, edvClient, kmsClient, profileSigner
-} = {}) {
-  const {hmac, keyAgreementKey} = edvConfig;
-
-  const doc = new EdvDocument({
-    invocationSigner: profileSigner,
-    id: docId,
-    keyAgreementKey: new KeyAgreementKey({
-      id: keyAgreementKey.id,
-      type: keyAgreementKey.type,
-      invocationSigner: profileSigner,
-      kmsClient
-    }),
-    hmac: new Hmac({
-      id: hmac.id,
-      type: hmac.type,
-      invocationSigner: profileSigner,
-      kmsClient
-    }),
-    client: edvClient
-  });
-  return doc.read();
-}
-
-async function _getProfileSigner({
-  kmsClient, profileAgentRecord, zcaps
-} = {}) {
-  const profileCapabilityInvocationKeyZcap =
-    zcaps.profileCapabilityInvocationKey;
-  const {data} = profileAgentRecord.secrets.seed;
-  const seed = Buffer.from(data);
-  const {capabilityAgent} = await getAgents({
-    profileAgent: profileAgentRecord.profileAgent,
-    secrets: {seed}
-  });
-  const profileAgentSigner = capabilityAgent.getSigner();
-  const profileSigner = await AsymmetricKey.fromCapability({
-    capability: profileCapabilityInvocationKeyZcap,
-    invocationSigner: profileAgentSigner,
-    kmsClient
-  });
-  return profileSigner;
 }
 
 /**
