+++ lib/utils.js
@@ -3,6 +3,10 @@
  */
 import * as bedrock from '@bedrock/core';
 import * as database from '@bedrock/mongodb';
+import * as profileAgents from './profileAgents.js';
+import {
+  AsymmetricKey, Hmac, KeyAgreementKey
+} from '@digitalbazaar/webkms-client';
 import assert from 'assert-plus';
 import {CapabilityAgent} from '@digitalbazaar/webkms-client';
 import {CryptoLD} from 'crypto-ld';
@@ -13,6 +17,7 @@ import {
 import {
   Ed25519VerificationKey2020
 } from '@digitalbazaar/ed25519-verification-key-2020';
+import {EdvDocument} from '@digitalbazaar/edv-client';
 import {getAppIdentity} from '@bedrock/app-identity';
 import {httpsAgent} from '@bedrock/https-agent';
 import {promisify} from 'node:util';
@@ -113,4 +118,55 @@ export function parseKeystoreId(keyId) {
     throw new Error(`Invalid key ID "${keyId}".`);
   }
   return keyId.slice(0, idx);
+}
+
+export async function getEdvConfig({edvClient, profileSigner} = {}) {
+  return edvClient.getConfig({
+    invocationSigner: profileSigner,
+  });
+}
+
+export async function getEdvDocument({
+  docId, edvConfig, edvClient, kmsClient, profileSigner
+} = {}) {
+  const {hmac, keyAgreementKey} = edvConfig;
+
+  const doc = new EdvDocument({
+    invocationSigner: profileSigner,
+    id: docId,
+    keyAgreementKey: new KeyAgreementKey({
+      id: keyAgreementKey.id,
+      type: keyAgreementKey.type,
+      invocationSigner: profileSigner,
+      kmsClient
+    }),
+    hmac: new Hmac({
+      id: hmac.id,
+      type: hmac.type,
+      invocationSigner: profileSigner,
+      kmsClient
+    }),
+    client: edvClient
+  });
+  return doc.read();
+}
+
+export async function getProfileSigner({
+  kmsClient, profileAgentRecord, zcaps
+} = {}) {
+  const profileCapabilityInvocationKeyZcap =
+    zcaps.profileCapabilityInvocationKey;
+  const {data} = profileAgentRecord.secrets.seed;
+  const seed = Buffer.from(data);
+  const {capabilityAgent} = await profileAgents.getAgents({
+    profileAgent: profileAgentRecord.profileAgent,
+    secrets: {seed}
+  });
+  const profileAgentSigner = capabilityAgent.getSigner();
+  const profileSigner = await AsymmetricKey.fromCapability({
+    capability: profileCapabilityInvocationKeyZcap,
+    invocationSigner: profileAgentSigner,
+    kmsClient
+  });
+  return profileSigner;
 }
